Backend/requirements.txt
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Pillow==10.0.1
config==0.5.1
django>=4.2,<5.3